+++ backend/src/main/java/com/cognizant/miam/configurations/HashingConfig.java
@@ -4,8 +4,5 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class HashingConfig {
-    /*@Bean
-    public PasswordEncoder encoder() {
-        return new StandardPasswordEncoder();
-    }*/
+
 }
